[FLINK-14441][table-api] Fix ValueLiteralExpression#getValueAs when value is Period
When value is Period, getValueAs(Integer.class) returns a Long value.

This closes #9931

+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/expressions/ValueLiteralExpression.java
@@ -106,7 +106,7 @@ public final class ValueLiteralExpression implements ResolvedExpression {
 
 		else if (valueClass == Period.class && clazz == Integer.class) {
 			final Period period = (Period) value;
-			convertedValue = period.toTotalMonths();
+			convertedValue = (int) period.toTotalMonths();
 		}
 
 		else if (valueClass == Integer.class && clazz == Period.class) {
